chore: move "handleWebSocketEvent" to "ws"
