+++ src/node/SetupServerCommonApi.ts
@@ -18,7 +18,7 @@ import type { WebSocketHandler } from '~/core/handlers/WebSocketHandler'
 import { mergeRight } from '~/core/utils/internal/mergeRight'
 import { devUtils } from '~/core/utils/internal/devUtils'
 import type { SetupServerCommon } from './glossary'
-import { handleWebSocketEvent } from '~/core/utils/handleWebSocketEvent'
+import { handleWebSocketEvent } from '~/core/ws/handleWebSocketEvent'
 import { webSocketInterceptor } from '~/core/ws/webSocketInterceptor'
 
 export const DEFAULT_LISTEN_OPTIONS: RequiredDeep<SharedOptions> = {
